requirements.txt
Flask>=2.0.0
werkzeug>=2.0.0
ffmpeg-python>=0.2.0